# [cli] lazy load imports to make --help show faster