--- a/packages/expo/cli/config/index.ts
+++ b/packages/expo/cli/config/index.ts
@@ -43,8 +43,12 @@ export const expoConfig: Command = async (argv) => {
   }
 
   // Load modules after the help prompt so `npx expo config -h` shows as fast as possible.
-  const { configAsync } = await import('./configAsync');
-  const { logCmdError } = await import('../utils/errors');
+  const [
+    // ./configAsync
+    { configAsync },
+    // ../utils/errors
+    { logCmdError },
+  ] = await Promise.all([import('./configAsync'), import('../utils/errors')]);
 
   return configAsync(getProjectRoot(args), {
     // Parsed options