--- a/packages/expo/cli/log.ts
+++ b/packages/expo/cli/log.ts
@@ -1,5 +1,3 @@
-import { EXPO_DEBUG } from './utils/env';
-
 export function time(label?: string): void {
   console.time(label);
 }
@@ -21,7 +19,7 @@ export function log(...message: string[]): void {
 }
 
 export function debug(...message: string[]): void {
-  if (EXPO_DEBUG) console.log(...message);
+  if (require('./utils/env').EXPO_DEBUG) console.log(...message);
 }
 
 /** Log a message and exit the current process. If the `code` is non-zero then `console.error` will be used instead of `console.log`. */