--- a/packages/expo/cli/prebuild/index.ts
+++ b/packages/expo/cli/prebuild/index.ts
@@ -50,6 +50,7 @@ export const expoPrebuild: Command = async (argv) => {
     );
   }
 
+  // Load modules after the help prompt so `npx expo prebuild -h` shows as fast as possible.
   const { prebuildAsync } = await import('./prebuildAsync');
   const { logCmdError } = await import('../utils/errors');
   const { resolvePlatformOption, resolveSkipDependencyUpdate } = await import('./resolveOptions');