# Why

Noticed `--help` flags were a bit sluggish. This PR improves times:
```
Before: npx expo prebuild -h  0.51s user 0.12s system 105% cpu 0.593 total
After: npx expo prebuild  -h  0.05s user 0.01s system 99% cpu 0.060 total
```

<!--
Please describe the motivation for this PR, and link to relevant GitHub issues, forums posts, or feature requests.
-->

# How

Lazily load command modules after the help command and argument parsing, this means invalid argument failures are faster too. This PR also makes the config commands slightly faster based on complexity.

# Test Plan

- `time npx expo prebuild -h`
- `time npx expo config -h`