@EvanBacon 2022-01-26
@ide I'm not sure of any certain way to prevent regressions. The bin file and all arg files (ex `packages/expo/cli/config/index.ts`) should just be considered pretty sensitive and not have any analytics or other massive imports added to them.