@ide 2022-01-26
A docblock that explains that in the code might be the best way forward, then, at least in the main index file.

If it becomes a problem we can use a bundler that ignores dynamic imports to estimate the size of the non-dynamic import graph.