@bbarthec 2022-01-26
Just an idea that might not be worth exploring at all: maybe we should generally switch to import things lazily everywhere in the cli 🤔 it seems that there're some solutions available (but having limitations as well), like https://github.com/sindresorhus/import-lazy 👀