@EvanBacon 2022-01-26
@bbarthec we use an auto lazy load babel plugin in expo-cli but it doesn't support modern techniques for minifying like ncc. The `import` syntax supports also supports types automatically which is pretty nice.